requirements.txt
torch==1.5.0
torchvision==0.6.0
numpy==1.18.3
matplotlib==3.2.1
Pillow==7.1.2
